ECP SSO
ECP allows non-browser clients to follow an SSO flow similar to browser
SSO but without redirects.

This IdP implementation of the ECP spec is not complete, but very close.
Notably we don't validate channel bindings sent by the ECP client or SP.
Shibboleth running as a SP, however, appears to not use channel bindings
in favor of using signing and/or encryption to secure messages against
tampering by the client.

We do validate the SP signature in the AuthnRequest relayed by the
client and we verify that the SP is registered.